Refactor results page to use Suspense and separate client-side component
